+++ PSNewScriptFile/changelog.md
@@ -2,17 +2,42 @@
 
 ## 1.0.1 (2020-07-27)
 
-- Upd: Changes to New-PSScript.ps1
-- Upd: Reordered parameters
-- Upd: Added all validation parameters that PowerShell uses
-- Upd: Fixed formatting violations in Changelog.md
+- New: strings.psd1 - Updated string table
 - New: Added new script parameters
 - New: configuration.properties.ps1 file
 - New: Add-ParametersToScriptfile.ps1 file
 - New: Script for building scripted template with all properties requested
-- Upd: Changes to Initialize-NewScript.ps1
 - New: Remoevd example (this is an internal function)
 - New: Added new code to check for properties to be added and removed for both parameter and attribute lists
+- New: Added new lists parameterList and validationAttributeList
+- New: Added code for seperating both lists for validation and properties
+- New: Added new Write-PSFMessages
+- New:  Find-ExistingScript.ps1
+- New: Added Stop-PSFunction -string output
+- New: Added entry point to  Find-ExistingScript
+- Upd: Changes to New-PSScript.ps1
+- Chg: Changed ParameterSet to ParameterSetName
+- Chg: Changed ValueFromRemainingArguements to ValueFromRemainingArguments
+- Chg: Changed EnableException [bool] to [switch]
+- Chg: Resorted and fixed begin block for $scriptParameters
+- Chg: configuration.properties.ps1
+- Chg: Changed ParameterSet to ParameterSetName
+- Chg: Changed ValueFromRemainingArguements to ValueFromRemainingArguments
+- Chg: Add-ParametersToScriptfile.ps1
+- Chg: Renames ParameterList to UserSuppliedParameterList
+- Chg: Changed ParameterSet to ParameterSetName
+- Chg: Changed ValueFromRemainingArguements to ValueFromRemainingArguments
+- Chg: Renamed Parameter to parameter
+- Chg: Changed ParameterSet to ParameterSetName
+- Chg: Removed unwanted code
+- Chg: Save-PSScriptFile.ps1
+- Chg: PSNewScriptFile.psd1
+- Upd: Reordered parameters
+- Upd: Added all validation parameters that PowerShell uses
+- Upd: Fixed formatting violations in Changelog.md
+- Upd: Changes to Initialize-NewScript.ps1
+- Upd: Initialize-NewScript.ps1
+- Upd: Incremented build to 1.0.0 to 1.0.1
 
 ## 1.0.0 (2020-07-26)